refactor(tests/e2e): extract reusable incoming/outgoing payment create interceptors
